Fix compilation warnings
TravisCi is reporting the following compilation warnings when compiling
the app in the test environment.

```
warning: the result of the expression is ignored (suppress the warning
by assigning the expression to the _ variable)
  lib/honeydew/success_mode/log.ex:7

warning: the result of the expression is ignored (suppress the warning
by assigning the expression to the _ variable)
  lib/honeydew/success_mode/log.ex:8
```

This is because we are purging all log statements lower than warning in
the test environment

```
config :logger
  compile_time_purge_level: :warn
```

This change puts both the calculations and text for the log message into
a function so they can all be purged together and not cause compilation
warnings.