Update typespec so args is a term instead of list (#27)
Since Honeydew.worker_spec/3 passes the args from the `module_and_args`
param directly to the Honeydew.Worker.init/1 callback, there is no
reason why the arguments here have to be a list. After double-checking
the args param for GenServer.start_link/3 and the associated
GenServer.init/1 callback, it appears common for the args param to be a
term instead of a list.