Prefer :capture_log tag when log data not tested (#34)
Since we are only using capture_log to silence the error output in the
tests, it seems like it is less intrusive to use the :capture_log tag
instead of wrapping everything in a function call. If a test fails
that's tagged with :capture_log, it will automatically dump all log
output to the console.