Add docs for async, yield, and filter (#36)
* Add for async, yield, and filter

This adds documentation and typespecs for the following functions:

  * `Honeydew.filter/2`
  * `Honeydew.async/3`
  * `Honeydew.yield/2`

It also adds a `Honeydew.Job.t` type and starts filling in the types
for the various fields on that struct. Finally, this fixes the typespec
on the `Honeydew.Queue.filter/2` callback. The previous callback said
that `filter/2` should return a non-empty list, which is not required.

* Additional doc improvements

* Fix docs for when queue doesn't exist