Failing test for FailureMode.Retry times: 1
I'm unclear about the meaning of the `times` argument. Does it mean try
this job x times, or try it the first time and then retry it x times if
it fails? I tried setting `times` to 1 and it didn't behave as expected,
so I thought I'd open an issue about it.